requirements.txt
imageio==2.8.0
numpy==1.19.0
Pillow==7.2.0
tensorflow==2.1.0
